howtos/trigger-cisco-config/requirements.txt
paramiko
requests
pyyaml
